Add files via upload
1. List Manipulaon: 
a. Write a funcon that takes a list of numbers and returns the sum of all even numbers in the list. 
b. Write a funcon that takes a list of words and returns a new list containing only the words that start with the leer 'A' or 'a'.
 c. Write a funcon that takes a list of characters and counts the number of occurrences of each character in the list. 
2. List Searching and Filtering: 
a. Write a funcon that takes a list of numbers and returns the index of the ﬁrst occurrence of a speciﬁc target number in the list. 
b. Write a funcon that takes a list of words and returns a new list containing only the words that have a length greater than or equal to a speciﬁed threshold. 
c. Write a funcon that takes a list of items and returns a new list containing only the items that sasfy a certain condion, such as being divisible by a speciﬁc number or having a speciﬁc property.
 3. List Operaons and Transformaons:
 a. Write a funcon that takes a list of numbers and reverses the order of the elements in the list.
 b. Write a funcon that takes a list of characters and converts all uppercase characters to lowercase. 
c. Write a funcon that takes a list of mixed numbers (whole numbers and fracons) and converts them to decimal form.